Cache DNS records based on the record requested
Currently Amp/DNS will only cache (in-memory) for the record type returned from
the DNS server. that means if we look up a DNAME record and instead receive a
CNAME record, it'll be cached under CNAME, leading amp to make additional
requests for the DNAME record it technically already received.

This commit changes this behavior and instead caches the record with the
requested type as the key rather than the returned type.